+++ cansocket/examples/send_errors.sh
@@ -1,0 +1,9 @@
+cansend vcan0 20000001#0000000000000000 # tx timeout
+cansend vcan0 20000002#2000000000000000 # lost arbitration
+cansend vcan0 20000004#0004000000000000 # controller error (rx overflow)
+cansend vcan0 20000008#0000000000000000 # protocol error (unspecified)
+cansend vcan0 20000010#0000000000000000 # transceiver error
+cansend vcan0 20000020#0000000000000000 # no ack on transmit
+cansend vcan0 20000040#0000000000000000 # bus-off condition
+cansend vcan0 20000080#0000000000000000 # bus-error condition
+cansend vcan0 20000100#0000000000000000 # restarted after bus-off