+++ cansocket/src/lib.rs
@@ -1,8 +1,10 @@
 #![doc = include_str!("../README.md")]
 
+mod error;
 mod frame;
 mod socket;
 
 pub use embedded_can::Id;
+pub use error::*;
 pub use frame::*;
 pub use socket::*;