+++ cansocket/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix `Socket` doc test
 - Added option to enable protocol errors
 - Added async example with smol
+- Added decoding error state indicator frames
 
 ## v0.1.0